+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [x] Native search
 - [x] Card screen
+- [x] Add “Lost Password” button on Login screen
 - [x] Support formatted authors names
 - [x] Improve loading loans images